Implementation use cases and business logic

+++ src/main/java/com/example/coupon_service/api/dto/ProductDto.java
@@ -3,13 +3,13 @@ package com.example.coupon_service.api.dto;
 import jakarta.validation.constraints.NotNull;
 
 public record ProductDto(
-        @NotNull String productId,
+                @NotNull String productId,
 
-        @NotNull String name,
+                @NotNull String name,
 
-        @NotNull String category,
+                @NotNull String category,
 
-        @NotNull Integer stock,
+                @NotNull Integer stock,
 
-        @NotNull Integer price) {
+                @NotNull Double price) {
 }
